+++ Contrib/at67/expression.cpp
@@ -410,7 +410,7 @@ namespace Expression
         bool inString = false;
         bool inComment = false;
 
-        for(size_t i=0; i<input.size(); i++)
+        for(int i=0; i<int(input.size()); i++)
         {
             // Check for string
             if(!inComment  &&  input[i] == '\"') inString = !inString;
